+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this crate will be documented in this file.
 
-## 0.1.0
+## 0.2.0 - 2024-09-26
+
+- Support alignments releative to the whole Ui
+- Support allocate no space & content row & content column
+
+## 0.1.0 - 2024-09-25
 
 - First release